hierarchy -> squad, set -> line
