+++ mod/badges/set/abstract/affinity_badge_squad.rb
@@ -1,34 +1,34 @@
 #! no set module
 
-# module AffinityBadgeHierarchy
-#   include BadgeHierarchy
+# module AffinityBadgeSquad
+#   include BadgeSquad
 #
 #   def hierarchy map
 #     @map = {}
-#     map.each do |action, badge_set|
-#       if badge_set.values.first.is_a? Hash
+#     map.each do |action, badge_line|
+#       if badge_line.values.first.is_a? Hash
 #         @map[action] = {}
-#         badge_set.each do |affinity, affinity_badge_set|
-#           @map[action][affinity] = Abstract::BadgeSet.new affinity_badge_set
+#         badge_line.each do |affinity, affinity_badge_line|
+#           @map[action][affinity] = Abstract::BadgeLine.new affinity_badge_line
 #         end
 #       else
-#         @map[action] = Abstract::BadgeSet.new badge_set
+#         @map[action] = Abstract::BadgeLine.new badge_line
 #       end
 #     end
 #   end
 #
 #   # returns a badge if the threshold is reached
 #   def earns_badge count, action, affinity_type=nil
-#     badge_set(action, affinity_type).earns_badge count
+#     badge_line(action, affinity_type).earns_badge count
 #   end
 #
 #   [:threshold, :level, :level_index].each do |method_name|
 #     define_method method_name do |action, affinity_type, badge_key|
-#       badge_set(action, affinity_type).send method_name, badge_key
+#       badge_line(action, affinity_type).send method_name, badge_key
 #     end
 #   end
 #
-#   def badge_set action, affinity_type
+#   def badge_line action, affinity_type
 #     validate_badge_args action, affinity_type
 #     affinity_type ? @map[action][affinity_type] : @map[action]
 #   end
@@ -37,7 +37,7 @@
 #     unless @map[action]
 #       raise StandardError, "not supported action: #{action}"
 #     end
-#     if affinity_type && !@map[action][affinity_type].is_a?(Abstract::BadgeSet)
+#     if affinity_type && !@map[action][affinity_type].is_a?(Abstract::BadgeLine)
 #       raise StandardError,
 #             "affinity type #{affinity_type} not supported for action #{action}"
 #     end
