--- mod/badges/set/type/project/badge_hierarchy.rb
@@ -1,13 +1,0 @@
-#! no set module
-
-class BadgeHierarchy
-  extend Abstract::BadgeHierarchy
-
-  add_badge_set :create,
-                project_launcher: 1,
-                &create_type_count(ProjectID)
-
-  add_badge_set :discuss,
-                project_q_a: [1, :silver],
-                &type_plus_right_edited_count(ProjectID, DiscussionID)
-end
