+++ mod/badges/set/type/user/badge_count.rb
@@ -2,7 +2,7 @@ include_set Abstract::Certificate
 
 def badge_count level=nil
   count = 0
-  Abstract::BadgeHierarchy::BADGE_TYPES.each do |badge_type|
+  Abstract::BadgeSquad::BADGE_TYPES.each do |badge_type|
     next unless (badge_pointer = field(badge_type, :badges_earned))
     count += badge_pointer.badge_count(level)
   end
